Added git commit command, which uses BBEdit as the editor
The script will output three lines into your shell worksheet. It can not execute these three lines directly as a script, because Git will wait for you to save and edit the file in BBEdit, and BBEdit will wait (patently, but with a modal dialog box) for Git to quit.

My solution is to do everything in a shell worksheet, which will not block BBEdit